rutabaga_gfx: fix race between CrossDomainWorker
talex5@ reports:

"CROSS_DOMAIN_CMD_GET_IMAGE_REQUIREMENTS writes to the shared ring
immediately, and thus races with any Wayland data being written to
it. To avoid this, I open the device file twice, using one for
images and the other for Wayland protocol data."

A solution is just to have them use different pieces of memory.
The guest already designates different virtio-gpu rings for both of
these.  We didn't seem to need separate rings before because the
host was slower due an unrelated polling thread.

BUG=b:259268477
TEST=run cross-domain, observe no crash between image requirements
     and wayland data

Change-Id: Ie8527302f2542a19ee09a88716d3553672600a3e