rutabaga_gfx: separate namespace for Wayland read pipes
talex5@ reports:

"When sending a pipe, the guest has to know what ID the host will
assign to it. The guest guesses, by adding 2 to the last ID it
knows about, and the host checks its guess. If it was wrong
(e.g. because the host created a new ID at the same time) then you
get a protocol error."

A solution is a separate namespace for Wayland read pipes.

BUG=b:259268477
TEST= - copy+paste still works
      - didn't repro exact case, but should just work (TM)

Change-Id: I76938b8b8105289bbfa4c5891eb95f7035d5196a