rutabaga_gfx: build.rs: use runtime feature checks
The #[cfg(feature = "...")] checks are evaluated when build.rs itself is
built. It is preferable to check the CARGO_FEATURE_* environment
variables when build.rs is running instead so it uses the up-to-date
configuration.

This also simplifies the code since it eliminates the need for many of
the cfg checks that were only necessary to prevent unused code warnings.

<https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-build-scripts>

Change-Id: I2b2d9e11d1292c27a7a028ed83a545aca4490573
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4860837
Reviewed-by: Frederick Mayle <fmayle@google.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>