rutabaga: add support for virgl's resource_map_fixed
Plumb virgl's resource_map_fixed into rutabaga to simplify BO mappings.
The original commit was based on the prev `virgl_renderer_resource_map2`
API, but the version that actually landed in virglrenderer is called
`virgl_renderer_resource_map_fixed` with a simpler signature that only
takes a resource handle and address pointer.

This commit is based on:
https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1374

Signed-off-by: Sergio Lopez <slp@redhat.com>
Signed-off-by: Dorinda Bassey <dbassey@redhat.com>