rutabaga_gfx: make vulkano feature functional
Uncomment the `vulkano` feature and the `vulkano` crate dependency to
make it possible to enable without editing Cargo.toml.

The feature is not in the default set, and the vulkano crate is optional
and only enabled when the feature is requested, so there should be no
change in the default build.

This fixes the build of crosvm with the top-level `vulkano` feature
enabled, making `cargo build --all-features` usable.

Change-Id: I27dcf60c954bd839b3b24ddb0454ca7ca2ab6923
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6064662
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Ryan Neph <ryanneph@google.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>