rutabaga_gfx: snapshot/restore rutabaga core to directory
... to avoid saving/loading snapshots to *both* a buffer and a
directory. This makes the rust part of rutabaga more consistent
with the ffi part which was just saving the buffer to a file
called "snapshot" in the directory. Users of rutabaga (virtio gpu
device) are responsible for packing/unpacking the directory if a
singular value is needed.

Bug: b/369615058
Test: cvd start \
      --gpu_mode=gfxstream_guest_angle_host_swiftshader \
      --enable_virtiofs=false
Test: cvd snapshot_take \
      --force \
      --auto_suspend \
      --snapshot_path=/tmp/snapshot1
Test: cvd create --snapshot_path=/tmp/snapshot1
Change-Id: Ie63271a89abb93d0c5c4eeb41a1a275ceca63e23
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6235987
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Jason Macnak <natsu@google.com>