rutabaga_gfx: don't use workspace dependencies
This allows rutabaga_gfx to explicitly manage it's own
dependencies, which is important since it aspires to packaged
across various distros and separable from crosvm.  It also
is a bit more selective on dependencies than the rest of crosvm,
so the likelihood of version skew is a bit lower.

BUG=419304200
TEST=none

Change-Id: Ia7d6013c3139c0ae518317ffe0161deafbf26c83
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6570569
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>