rutabaga_gfx: add Magma support
This adds stub support for Magma context types.  For the
projected benefits of Magma, please see the Mesa3D RFC:

https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/33190

I think we'll want to merge some aspects of it in upstream
Rutabaga at-least to cut down on the patch stack.  Rutabaga
allows experimental code, but it's subject to deletion if
the plots/schemes associated with experimental doesn't
work out.

TEST=cargo build --package kumquat_virtio
     ./target/debug/kumquat --capset-names="magma:"