rutabaga_gfx: remove guest_cpu_mappable from Resource3DInfo
This was questionable change made to accomodate virtio-video for
CrOS.

virtio-video is being deprecated in favor of virtio-media,
but virtio-media might take another six months to upstream.
And no one from Google is working on it.  And even if it's
upstreamed, it's unlikely CrOS will switch to it.

The best rutabaga can do is decrease the public API exposure,
with a getter function rather than plumbing into an more
generic struct.