rutabaga_gfx: v0.1.71 release
This release is sufficient to remove rutabaga_gfx from
crosvm.

In the past, a crates.io upload was just done to unblock
something or even test a change against a CI system (which is
built around crates.io).  A crates.io upload was not accompanied
with a Github tag.

Now, given the super-serious and totally-not-a-joke nature
of rutabaga, we should actually have a 1:1 to mapping of
crates.io upload to a Github tag for the rutabaga_gfx
and mesa3d_util crates.  To test changes, we should use
pre-releases.  A release will only be created after a
pre-release crate passes a downstream CI (like crosvm-based).

The Github tag will also apply to the rutabaga_gfx_ffi crate.
There isn't neccessarily a crates.io upload associated with
every Github tag for rutabaga_gfx_ffi.  This is because
everyone who distro-packages it doesn't use crates.io.

A note about distro-packaging.  The requirements for
distro-packaging are more strict than crates.io.  For example,
while it's fine to create your mesa3d_util crate and upload
it, would that be acceptable for a distro?  Probably not.

For mesa3d_util, the plan is a mesa3d::util one day anyways.
So if intend to distro package mesa3d_util, I suggest waiting
for the mesa3d crate instead.  Or accelerating development
of the mesa3d crate.

The use of mesonbuild can actually helps these issues.

The current status is good enough to unlock the crosvm migration,
but use at your risk elsewhere.