rutagaba_gfx: Add support for specifying drm fd for virglrenderer
The server now supports multiple graphics devices, so it is beneficial
to add the option to specify which GPU to use.

The patch allows users to provide a GPU path for virglrenderer. Due to
the virglrenderer limit, only one GPU is used.

In order to support receving a set of GPU paths, some interfaces have
been updated, including the renaming of `RutabagaChannel` to
`RutabagaPath`. The FFI component will not be updated due to
compatibility.

Users can add a path to `RutabagaPaths` using the `set_rutabaga_paths()`
method provided by `RutabagaBuilder`. On Linux, a typical path could be:

```rust
RutabagaPath {
    path: PathBuf::from("/dev/dri/renderD128"),
    path_type: RUTABAGA_PATH_TYPE_GPU
}
```

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>