rutabaga_gfx: Fix a path-matching issue and improve logs
`PathBuf::starts_with()` semantically differs from `String`'s.
The `is_valid_gpu_path()` checks if the GPU path matches the format of
`/dev/dri/renderD{X}`, rather than checking its parent directory name.
So we have to convert the `Pathbuf` to a `String` firstly.

When no GPU path is available, sometimes it is expected (the user didn't
feed one) while sometimes not (the user feed but the path is not
existing), print a log to clarity the situation.

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>