RutabagaHandle as Either<MesaHandle, AhbInfo>
To support AHB as rutabaga resource, extend RutabagaHandle to an enum of
MesaHandle or AhbInfo. MesaHandle isn't enough to store every
information which AHB needs.

And RutabagaHandle(especially AhbInfo) is expected to be used for blob,
not fence, so make blob-related function use RutabagaHandle instead of
MesaHandle.

It doesn't have behavior change yet.

(refer to ag/36824111)

Bug: 453989233
Test: build (there is no behavior change yet)